requirements.txt
astroid==3.2.4
certifi==2024.8.30
charset-normalizer==3.3.2
dill==0.3.8
idna==3.8
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.6
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
tomlkit==0.13.2
urllib3==2.2.2
